+++ src/Model/Analytic/AnalyticResource.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace ApiVideo\Client\Model\Analytic;
+
+class AnalyticResource
+{
+    /** @var string */
+    public $type;
+
+    /** @var string */
+    public $id;
+}