+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 All changes to this project will be documented in this file.
 
+## [1.2.0] - 2019-07-18
+### Added
+- Improve Documentation
+- Light Refactoring to support new Analytics response
+- Add Analytics Session Events support
+
+## [1.1.0] - 2019-06-13
+### Added
+- Constructors for production and sandbox environments
+
 ## [1.0.0] - 2019-03-06
 ### Added
 - New resource /account to get quota and term informations.